# Make the `decodeHuffman` function, in `src/core/jpg.js`, slightly more efficient